pyproject.toml
[project]
name = "local-perplexity"
version = "0.1.0"
description = ""
authors = [
    {name = "Kael Ducatti",email = "contato.mikaelducatti@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "streamlit (>=1.45.1,<2.0.0)",
    "langgraph (>=0.4.8,<0.5.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "ruff (>=0.11.13,<0.12.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "tavily-python (>=0.7.5,<0.8.0)",
    "ipython (>=9.3.0,<10.0.0)"
]

[tool.poetry]
packages = [{include = "local_perplexity", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
